requirements.txt
# requirements.txt
# Core trading dependencies
python-binance==1.0.19
websockets==12.0
numpy>=1.24.0
pandas>=2.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
asyncpg>=0.28.0  # For PostgreSQL support

# Configuration and environment
python-dotenv>=1.0.0
pydantic>=2.0.0

# Telegram bot (optional, for later phases)
aiogram>=3.0.0

# Utilities
colorama>=0.4.6  # For colored console output
cryptography>=41.0.0  # For secure operations

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0  # Code formatting
mypy>=1.5.0   # Type checking

# Optional: Advanced technical analysis
# ta-lib>=0.4.25  # Uncomment if you want TA-Lib instead of custom indicators

# HTTP client with connection pooling
aiohttp>=3.9.0